backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pymongo==4.6.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
bcrypt==4.1.2
pydantic==2.5.1
python-dotenv==1.0.0
httpx==0.25.2
emergentintegrations --extra-index-url https://d33sy5i8bnduwe.cloudfront.net/simple/